replica/requirements.txt
Django==3.0.4
psycopg2-binary==2.8.4
psycopg2==2.8.4
pytz
sqlparse
jsonfield
PyJWT==1.7.1
cryptography==2.8
python-dotenv==0.10.1
social-auth-app-django==3.1.0
djangorestframework==3.9.0
pymongo==3.7.2
python-dotenv==0.9.1
